Refactor rewards chart panel so it accepts loading and error states
